pyproject.toml
[project]
name = "kuva_mlops"
version = "0.1.0"
description = "This project contains all the automations needed to perform MLOps"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "encord>=0.1.170",
    "google-cloud-storage>=3.1.1",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy==1.26.4",
    "pillow==10.4.0",
    "pipeline",
    "pre-commit>=4.2.0",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "ruff>=0.12.5",
]

[build-system]
requires      = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where    = ["."]

[tool.uv.sources]
pipeline = { git = "https://github.com/traktorspace/spacelab-pipeline.git", branch = "develop" }

[tool.lint.ruff]
select = [ 
    "E",
    "F",
    "A",
    "DTZ",
    "NPY",
    "I",
    "ISC",
    "B003",
    "B004",
    "B015",
    "PTH",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D200",
    "W191",
    "W291",
    "W293",
    "N801",
    "N804",
    "N805",
    "T100",
    "S105",
    "S106",
    "S108",
    "S604",
    "S602",
    "S609",
    "UP003",
    "UP005",
    "UP006",
    "UP007",
    "UP008",
    "UP032",
    "UP035",
    "RUF001",
    "RUF200",
    "RUF013",
    "C901",
    "COM818",
    "RSE102",
    "EM101",
    "F841",
    "E501",
    ]
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".nox",
    ".pytype",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "venv",
    "__pycache__",
]

[tool.ruff]
line-length = 80 
unsafe-fixes = true



[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = false
docstring-code-line-length = 20

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I", "ISC"]
